// website/WebContent/src/app/generated/api/service-requester/model/serviceReputation.ts
/**
 * Service Requester REST Interface
 * Rest interface to communicate with the service requester.
 *
 * OpenAPI spec version: 0.0.2
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface ServiceReputation { 
    date?: number;
    other?: number;
    overall?: number;
    performance?: number;
    reputationMessage?: string;
    security?: number;
    usability?: number;
}