// website/WebContent/src/app/generated/api/service-requester/model/itemAndRequest.ts
/**
 * Service Requester REST Interface
 * Rest interface to communicate with the service requester.
 *
 * OpenAPI spec version: 0.0.2
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Offer } from './offer';
import { ReactNonInteractivePolicyProof } from './reactNonInteractivePolicyProof';


export interface ItemAndRequest { 
    acceptedOffer?: Offer;
    accessCredential?: ReactNonInteractivePolicyProof;
    requestName?: string;
    requestUUID?: string;
    serviceLink?: string;
    serviceState?: ItemAndRequest.ServiceStateEnum;
}
export namespace ItemAndRequest {
    export type ServiceStateEnum = 'BOOTING' | 'RUNNING';
    export const ServiceStateEnum = {
        BOOTING: 'BOOTING' as ServiceStateEnum,
        RUNNING: 'RUNNING' as ServiceStateEnum
    };
}